Use boto3 insted aws-glue-schema-registry, refactor classes and cleanup
