+++ fluvii/auth/config.py
@@ -3,7 +3,7 @@ import requests
 import time
 
 
-class AWSRegistryConfig(KafkaConfigBase):
+class GlueRegistryClientConfig(KafkaConfigBase):
     def __init__(self, aws_access_key_id, aws_secret_access_key, region_name, registry_name):
         self.aws_access_key_id = aws_access_key_id
         self.aws_secret_access_key = aws_secret_access_key
