+++ fluvii/consumer/consumer.py
@@ -1,5 +1,5 @@
 from confluent_kafka import DeserializingConsumer, TopicPartition
-from fluvii.avro import FluviiAvroDeSerializer
+from fluvii.avro import GlueAvroDeSerializer
 from fluvii.exceptions import NoMessageError, ConsumeMessageError, FinishedTransactionBatch, TransactionNotRequired
 from fluvii.general_utils import parse_headers, get_guid_from_message
 from copy import deepcopy
@@ -65,8 +65,8 @@ class Consumer:
             "partition.assignment.strategy": 'cooperative-sticky',
 
             # Registry Serialization Settings
-            "key.deserializer": FluviiAvroDeSerializer(self._schema_registry, schema_str='{"type": "string"}'),
-            "value.deserializer": FluviiAvroDeSerializer(self._schema_registry) if self._schema_registry else None,
+            "key.deserializer": GlueAvroDeSerializer(self._schema_registry, schema_str='{"type": "string"}'),
+            "value.deserializer": GlueAvroDeSerializer(self._schema_registry) if self._schema_registry else None,
         }
 
         if self._settings:
