+++ fluvii/schema_registry/glue_schema_registry.py
@@ -1,17 +1,20 @@
 import logging
-import mmh3
-import boto3
 import uuid
+
+import boto3
 from botocore.exceptions import BotoCoreError, ClientError
 from confluent_kafka.schema_registry import RegisteredSchema, Schema
+
 LOGGER = logging.getLogger(__name__)
 
-def getUUIDtoInt(schemaId) :
-    schema_id = uuid.UUID(schemaId)
-    return schema_id.int
 
-def getInttoUUID(schemaInt) : 
-    return uuid.UUID(int=schemaInt)
+def get_uuid_to_int(schema_id):
+    return uuid.UUID(schema_id).int
+
+
+def get_int_to_uuid(schema_int):
+    return uuid.UUID(int=schema_int)
+
 
 class EntityNotFoundException(Exception):
     def __init__(self, entity_type, entity_name):
@@ -20,12 +23,9 @@ class EntityNotFoundException(Exception):
         message = f"{entity_type} '{entity_name}' not found"
         super().__init__(message)
 
+
 class GlueSchemaRegistryClient:
     def __init__(self, auth_config=None, auto_init=True):
-        """
-        Initialize the Glue Schema Registry client.
-        """
-        
         self._auth = auth_config
         if auto_init:
             self._init_registry()
@@ -43,32 +43,15 @@ class GlueSchemaRegistryClient:
         return registry_id
 
     def register_schema(self, schema_name, schema, normalize_schemas=False):
-        """
-        Registers a schema under ``subject_name``.
-
-        Args:
-            subject_name (str): subject to register a schema under
-
-            schema (Schema): Schema instance to register
-
-        Returns:
-            string: Schema id
-
-        Raises:
-            SchemaRegistryError: if Schema violates this subject's
-                Compatibility policy or is otherwise invalid.
-
-        """  # noqa: E501
         try:
             schemas = self.client.list_schema_versions(SchemaId={
                     'RegistryName': self._auth.registry_name, 
                     'SchemaName': schema_name
             })
             schema = schemas['Schemas'][0]
-            return getUUIDtoInt(schema['SchemaVersionId'])
+            return get_uuid_to_int(schema['SchemaVersionId'])
         except self.client.exceptions.EntityNotFoundException:
             try:
-                #Check if schema already exists first
                 schema = self.client.create_schema(
                     RegistryId={
                         'RegistryName': self._auth.registry_name,
@@ -78,100 +61,49 @@ class GlueSchemaRegistryClient:
                     SchemaDefinition=schema.schema_str,
                     Compatibility='BACKWARD'
                 )
-                return getUUIDtoInt(schema['SchemaVersionId'])
+                return get_uuid_to_int(schema['SchemaVersionId'])
             except (BotoCoreError, ClientError) as e:
-                print(e)
                 LOGGER.debug(f"An error occurred while creating schema: {e}")
                 return None
         
     def get_schema(self, schema_id) :
-        schema_id_decode = str(getInttoUUID(schema_id));
+        schema_id_decode = str(get_int_to_uuid(schema_id));
         schema = self.client.get_schema_version(SchemaVersionId=schema_id_decode)
         return Schema(schema_str=schema['SchemaDefinition'],schema_type='AVRO')
 
+    # ToDO: Fix schema.int
+    # def get_latest_version(self, schema_name, schema_version_id, registry_name):
+    #     try:
+    #         response = self.client.get_schema_version(
+    #             SchemaId={
+    #                 'RegistryName': registry_name,
+    #                 'SchemaName': schema_name
+    #             },
+    #             SchemaVersionNumber={
+    #                 'LatestVersion': True
+    #             }
+    #         )
+    #         schema_id = uuid.UUID(response['SchemaVersionId'])
+    #         return RegisteredSchema(schema.int, response['SchemaDefinition'], registry_name, response['VersionNumber'])
+    #     except (BotoCoreError, ClientError) as e:
+    #         LOGGER.debug(f"An error occurred while retrieving schema version: {e}")
+    #         return None
 
-    def get_latest_version(self, schema_name, schema_version_id, registry_name):
-        """
-        Get a specific version of a schema from the Glue Schema Registry.
-
-        Args:
-            schema_name (str): Name of the schema.
-            schema_version_id (str): ID of the schema version.
-
-        Returns:
-            dict: The schema version information.
-
-        Raises:
-            BotoCoreError: If a Boto Core error occurs.
-            ClientError: If an error occurs in the AWS Glue client.
-        """
-        try:
-            response = self.client.get_schema_version(
-                SchemaId={
-                    'RegistryName': registry_name, 
-                    'SchemaName': schema_name
-                },
-                #SchemaVersionId=schema_version_id, 
-                SchemaVersionNumber={
-                    'LatestVersion': True
-                }
-            )
-            schema_id = uuid.UUID(response['SchemaVersionId'])
-            return RegisteredSchema(schema.int, response['SchemaDefinition'], registry_name, response['VersionNumber']);
-
-        except (BotoCoreError, ClientError) as e:
-            LOGGER.debug(f"An error occurred while retrieving schema version: {e}")
-            return None
-        
-    def update_schema_definition(self, schema_name, schema_definition): 
-        """
-        Update an existing schema in the Glue Schema Registry.
-
-        Args:
-            schema_name (str): Name of the schema.
-            schema_definition (dict): Definition of the updated schema.
-            schema_version_id (str): ID of the schema version to update.
-
-        Returns:
-            str: The version ID of the updated schema.
-
-        Raises:
-            BotoCoreError: If a Boto Core error occurs.
-            ClientError: If an error occurs in the AWS Glue client.
-        """
+    def update_schema_definition(self, schema_name, schema_definition):
         try:
             response = self.client.register_schema_version(
                 SchemaId={
                     'SchemaName': schema_name,
                     'RegistryName': self._auth.registry_name
                 },
-                #SchemaVersionId=schema_version_id,
                 SchemaDefinition=schema_definition
             )
             return response
-            return response['SchemaVersionId']
         except (BotoCoreError, ClientError) as e:
             LOGGER.debug(f"An error occurred while updating schema: {e}")
             return e
-            return None 
-
 
     def update_schema(self, schema_name, schema_definition, schema_version_id):
-        """
-        Update an existing schema in the Glue Schema Registry.
-
-        Args:
-            schema_name (str): Name of the schema.
-            schema_definition (dict): Definition of the updated schema.
-            schema_version_id (str): ID of the schema version to update.
-
-        Returns:
-            str: The version ID of the updated schema.
-
-        Raises:
-            BotoCoreError: If a Boto Core error occurs.
-            ClientError: If an error occurs in the AWS Glue client.
-        """
         try:
             response = self.client.update_schema(
                 SchemaId={
@@ -184,19 +116,8 @@ class GlueSchemaRegistryClient:
         except (BotoCoreError, ClientError) as e:
             LOGGER.debug(f"An error occurred while updating schema: {e}")
             return e
-            return None
 
     def delete_schema(self, schema_name):
-        """
-        Delete a schema from the Glue Schema Registry.
-
-        Args:
-            schema_name (str): Name of the schema.
-
-        Raises:
-            BotoCoreError: If a Boto Core error occurs.
-            ClientError: If an error occurs in the AWS Glue client.
-        """
         try:
             self.client.delete_schema(
                 SchemaId={
