+++ fluvii/schema_registry/schema_registry.py
@@ -1,14 +1,27 @@
-import boto3
-from aws_schema_registry import SchemaRegistryClient
+from confluent_kafka.schema_registry import SchemaRegistryClient
+from urllib.parse import urlparse, quote
 
+
+# fixes a bug in confluent-kafka TODO: keep a look out for this in >1.8.2, should be fixed since I took it from a MR.
+def patched_schema_loads(schema_str):
+    from confluent_kafka.schema_registry.avro import Schema
+    schema_str = schema_str.strip()
+    if schema_str[0] != "{" and schema_str[0] != "[":
+        schema_str = '{"type":' + schema_str + '}'
+    return Schema(schema_str, schema_type='AVRO')
+
+
+import confluent_kafka
+confluent_kafka.schema_registry.avro._schema_loads = patched_schema_loads
 import logging
 
 LOGGER = logging.getLogger(__name__)
 
 
 class SchemaRegistry:
-    def __init__(self, auth_config=None, auto_init=True):
+    def __init__(self, url, auth_config=None, auto_init=True):
         self.registry = None
+        self.url = url
         self._auth = auth_config
         if auto_init:
             self._init_registry()
@@ -21,11 +34,19 @@ class SchemaRegistry:
             return self.registry.__getattribute__(attr)
 
     def _init_registry(self):
-        # Pass your AWS credentials or profile information here
-        session = boto3.Session(aws_access_key_id=self._auth.aws_access_key_id,
-                                aws_secret_access_key=self._auth.aws_secret_access_key,
-                                region_name=self._auth.region_name)
-        glue_client = session.client('glue')
-        self.registry = SchemaRegistryClient(glue_client,
-                                             registry_name=self._auth.registry_name)
+        url = urlparse(self.url)
+        if self._auth:
+            auth = f"{quote(self._auth.username)}:{quote(self._auth.password)}@"
+        else:
+            auth = ''
+        scheme = url.scheme
+        if scheme:
+            url = url._replace(path=url.netloc, netloc='')
+        else:
+            if self._auth:
+                scheme = 'https'
+            else:
+                scheme = 'http'
+        url = url._replace(scheme='')
+        self.registry = SchemaRegistryClient({'url': f'{scheme}://{auth}{url.geturl()}'})
         LOGGER.info('Registry client initialized successfully!')
